+++ src/main.js
@@ -1,8 +1,11 @@
+import { getAuth, onAuthStateChanged } from 'https://www.gstatic.com/firebasejs/9.9.4/firebase-auth.js';
 import { Welcome } from './components/Welcome.js';
 import { Login } from './components/Login.js';
 import { Register } from './components/Register.js';
 import { Wall } from './components/wall.js';
-import { userObserver } from './firebase/auth.js';
+
+
+const auth = getAuth();
 
 const root = document.getElementById('root');
 const routes = {
@@ -28,16 +31,11 @@ window.onpopstate = () => {
   root.removeChild(root.firstChild);
   root.append(component());
 };
-
-root.appendChild(component());
-
-userObserver ((user)=> {
-  console.log (user)
+onAuthStateChanged(auth, (user) => {
   if (user) {
-    
- onNavigate ('/wall')
+    onNavigate('/wall');
   } else {
-    onNavigate ('/')
+    onNavigate('/');
   }
-
 });
+root.appendChild(component());
